Add files via upload
James Bond is the sales manager for a durable product whose price fluctuates weekly. The product
is advertised on three channels - print, radio and internet. Mr. Bond is interested in understanding which
advertising channel (print, radio and internet) is most effective.
The data consists of a single file called SalesData.csv. This is a CSV data file which consists of the
following fields - SALES, PRICE, PRINTADVT, RADIOADVT and INTERNETADVT for 300 weeks.
The SALES and PRICE variables record sales in ’000 units and price in dollars for each week
respectively. The three ADVT variables give the weekly expenditure in ’000 dollars for the specific
channel in that week.